pyproject.toml
[tool.poetry]
name = "recsys-challenge-2023"
version = "0.1.0"
description = "RecSys challenge 2023"
authors = ["Paolo Basso"]
readme = "README.md"
packages = [
    {include = "utils"},
]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
scipy = "^1.10.1"
pandas = "1.5.3"
matplotlib = "^3.7.1"
black = "^23.3.0"
mypy = "^1.1.1"
ipykernel = "^6.22.0"
pyarrow = "^11.0.0"
tqdm = "^4.65.0"
seaborn = "^0.12.2"
lightgbm = "^3.3.5"
kmodes = "^0.12.2"
numpy = "~1.23"
tensorflow-io-gcs-filesystem = "0.31.0"
ruff = "^0.0.261"
optuna = "3.1.1"
iguanas = "^0.1.4"
pyspark = "^3.4.0"
tensorflow = "^2.12.0"
iprogress = "^0.4"
ipywidgets = "^8.0.6"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.6"
imblearn = "^0.0"
telegram = "^0.0.1"
python-telegram-bot = "^20.3"
boto3 = "^1.26.136"
catboost = "^1.2"
xgboost = "^1.7.5"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
exclude = '''
(
    \.DS_Store |
    \.githooks |
    \.idea |
    \.poetry |
    \.pyc$ |
    \.vscode |
    data |
    \.venv |
    \.mypy_cache |
    \.git
)
'''

[tool.mypy]
python_version = 3.10
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
strict_optional = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true
check_untyped_defs = true
strict = true
disallow_subclassing_any = true
disallow_untyped_calls = true
exclude = [
    '.venv/',
    'venv/',
    '.git/',
    '.vscode/',
    '.idea/',
    'data',
]

[[tool.mypy.overrides]]
module = [
    "torch.*",
    "scipy.*",
    "sklearn.*",
    "kmodes.*",
    "pandas.*",
    "seaborn.*",
    "matplotlib.*",
    "category_encoders.*",
]
ignore_missing_imports = true

[tool.ruff]
select = ["E", "F", "W", "I", "NPY", "PD"]
ignore = []
exclude = [
    '.venv/',
    'venv/',
    '.git/',
    '.vscode/',
    '.idea/',
    'data',
]
line-length = 100
target-version = "py310"

[tool.ruff.pydocstyle]
convention = "google"